Added Softball + Baseball Updates
Added support for Softball with the All Sport CG *ONLY* in manual mode with game mode baseball selected. The softball program will not work correctly when connected to an All Sport 5000. Please use the Baseball program if you're All Sport CG is connected to an All Sport 5000. It's also worth noting that the softball set should work for baseball games if you're unable to connect to an All Sport 5000 as well. (For example, road games)

Corrected some spacing logic in Baseball for innings with multiple characters, and set TOP and BOT to be fully capitalized.